+++ src/crepe/api/Sprite.cpp
@@ -1,3 +1,4 @@
+#include <cmath>
 #include <memory>
 
 #include "../util/Log.h"
@@ -16,7 +17,7 @@ Sprite::Sprite(game_object_id_t id, const shared_ptr<Texture> image, const Color
 	  color(color),
 	  flip(flip),
 	  sprite_image(image),
-	  aspect_ratio(sprite_image->get_width() / sprite_image->get_height()) {
+	  aspect_ratio(static_cast<double>(sprite_image->get_width()) / sprite_image->get_height()) {
 	dbg_trace();
 
 	this->sprite_rect.w = sprite_image->get_width();
